requirements.txt
discord.py
python-dotenv
asyncpg
google-generativeai
flask